[Author: nigeltao]
For Safari/Mac drag-and-drop, use the dragginSequenceNumber from
the NSDraggingInfo to determine when to reset our cached contents
of the NSDraggingInfo's pasteboard.

This (the sequence number) appears to be more reliable than
resetting on the draggingExited or performDragOperation message,
since neither of those messages are received if, for example,
the web pages calls window.alert during a dragover handler,
since presumably the change of focus inadvertently kills the
drag-and-drop in progress.

PRESUBMIT=passed
R=playmobil
CC=gears-eng@googlegroups.com
APPROVED=playmobil
DELTA=18  (13 added, 2 deleted, 3 changed)
OCL=10550638
SCL=10566207