[Author: nigeltao]
First cut of on-screen canvas rendering.

The model is that a GearsCanvas remains an off-screen object, but it
now can have a DOM element peer (provided that the GearsCanvas isn't
in a worker thread) that can participate in the DOM tree, and knows
how to render the GearsCanvas' contents (i.e. the backing SkBitmap).

On IE, this uses an IHTMLPainter behavior and GDI+ to blit the
SkBitmap directly. Compared to the 'indirect' method of rendering a
GearsCanvas (i.e. calling encode() to create a PNG Blob to capture
(via LocalServer) to point an <img>'s src to), this actually renders
transparency properly on IE6.

On Firefox, I am not aware of an equivalent mechanism to IHTMLPainter,
and so the rendering peer is a (Firefox's native) <canvas> element,
and the data is transferred from the GearsCanvas gc to the native
canvas nc via nc.putImageData(gc.getImageData()). This is undoubtedly
inefficient, but it is correct, and we can do something faster (but
more complicated) later. Besides, Firefox already has a native canvas,
and so the Gears canvas' on-screen rendering performance on FF is not
as high a priority as the speed on IE (which has no native canvas).

I've punted on measuring rendering performance, since a call to
invalidateRenderingElement doesn't actually paint the pixels, it
schedules a paint to happen some time in the future, and therefore
measuring the clock time elapsed before and after an
invalidateRenderingElement call won't capture the total time to
draw the off-screen bitmap on-screen.

In order to avoid a conversion from SkBitmap's packing order and
GDI+'s packing order, Skia's packing was configured to be ARGB
(instead of Skia's default ABGR) on Win32 Gears builds.

This is very much pre-alpha quality (and has only been tested on
IE/Win32, Firefox/Win32, and very briefly on Firefox/Linux). The
API is also up for debate (e.g. should invalidation be automatic
or explicit). Nonetheless, it works, so I'm putting this out as a
first cut.

R=noel,reed
CC=gears-eng@googlegroups.com
APPROVED=noel
DELTA=611  (573 added, 22 deleted, 16 changed)
OCL=10562599
SCL=10670392