[Author: nigeltao]
Build fixes for 64-bit Ubuntu Hardy (GCC 4.2).

This builds a 64-bit Firefox/Linux Gears installer for Firefox3
only, not Firefox2, since it is simpler to do so, Firefox2 is
approaching end-of-life, and (IIUC) Firefox2 is not part of the
default Ubuntu Hardy install.

64-bittedness versus 32-bittedness is determined by the result
of "gcc -dumpmachine". For example, if you are running a 64-bit
kernel but a 32-bit userspace (including a 32-bit gcc) then
calling make will build the 32-bit Gears installer.

Also, the resultant 64-bit .xpi filename has an explicit ARCH
(i.e. the "-x86_64" in gears-linux-x86_64-dbg-1.2.3.4.xpi).
If building a 32-bit installer, there is no ARCH in the filename,
the same as before.

The binary libraries (e.g. libxpcom.so, libxul.so) that are new
to third_party are copied straight from Ubuntu's /usr/lib64/firefox.

AsyncTask's HandleEvent method was renamed to HandleAsyncTaskEvent,
in order to disambiguate it from the other HandleEvent names in
the source code (there are at least three -- there's a Gears concept
for JSEVENT_UNLOAD, there's the DOM's HandleEvent used in the
NPAPI port, and there's the Gecko event system used in the
Firefox port).

The second parameter to AsyncTask::HandleAsyncTaskEvent was
clarified to be an int, rather than either an int or a void*.
The latter were equivalent on a 32-bit system, but are obviously
not so on a 64-bit system. On inspection, the only uses of the
param arg was either unused (zero) or as an index to an array.

Blob- and file- related changes make casts from size_t to int64
explicit. On 32-bit systems, the cast loses no information. On
64-bit systems, size_t is unsigned and int64 is signed, and so
an implicit cast causes a compiler warning (and we compile with
-Wall -Werror).

BufferBlobs are now restricted to have a maximum length of
kint32max, or 2^31-1, regardless of whether or not the
system is 32-bit or 64-bit.

The assertion in i18n_strings.cc was removed because it caused
an unused code warning, since the compiler could determine that
the I18NStringKey enum was always <= the ARRAYSIZE(kLocalStrings).

The global std::vector data variable in blob_input_stream_ff_test.cc
was simply unused (and hence causing a compiler warning).

The sqlite fts?_porter.c change was to silence a compiler warning
where the naked a || b; expression was considered unused.

All tests (http://localhost:8001/tester/gui.html) are green, on
Firefox3 / Linux (Ubuntu Hardy x86_64).

PRESUBMIT=passed
R=cprince,dsymonds
CC=gears-eng@googlegroups.com
APPROVED=cprince,dsymonds
DELTA=216  (70 added, 12 deleted, 134 changed)
OCL=10624727
SCL=10670747