[Author: nigeltao]
Unbreak Firefox2/Mac.

This was broken by OCL 10624727 "Build fixes for 64-bit Ubuntu
Hardy (GCC 4.2)". That CL assumed that if $ARCH was not "x86_64",
then it was "i386", but on the Mac, it is "i386+ppc".

PRESUBMIT=passed
R=cprince
CC=gears-eng@googlegroups.com
APPROVED=cprince
DELTA=1  (0 added, 0 deleted, 1 changed)
OCL=10876769
SCL=10887485